python-services/data-ingestion/requirements.txt
# ParleyApp Data Ingestion Service Requirements
# Phase 1 Implementation

# Core dependencies
httpx==0.27.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0

# Async support
asyncio==3.4.3
aiofiles==23.2.1
apscheduler==3.10.4

# Data processing
pandas==2.2.0
numpy==1.26.3
pydantic==2.5.3

# MLB Data API
MLB-StatsAPI==1.7.0

# Retry and resilience
backoff==2.2.1
tenacity==8.2.3

# Monitoring and logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# API client utilities
python-dateutil==2.8.2
pytz==2023.3 